Introduction to Artificial Intelligent/Project/IT3160E-main/IT3160E-main/requirements.txt
tensorflow==2.12.0
matplotlib>=3.3.0
pillow>=9.0.0
pandas>=1.1.0
